+++ example/integration_test/bdk_full_cycle_test.dart
@@ -5,10 +5,11 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:payjoin_flutter/common.dart' as common;
+import 'package:payjoin_flutter/receive/v1.dart' as v1;
+import 'package:payjoin_flutter/send.dart' as send;
 import 'package:payjoin_flutter/uri.dart' as pay_join_uri;
 import 'package:payjoin_flutter_example/bdk_client.dart';
 import 'package:payjoin_flutter_example/btc_client.dart';
-import 'package:payjoin_flutter_example/payjoin_library.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
@@ -16,51 +17,83 @@ void main() {
   group('v1_to_v1', () {
     setUp(() async {});
     testWidgets('full_cycle', (WidgetTester tester) async {
-      final payJoinLib = PayJoinLibrary();
       final btcClient = BtcClient("sender");
       await btcClient.loadWallet();
       final sender = BdkClient(
-          "puppy interest whip tonight dad never sudden response push zone pig patch");
+          "wpkh(tprv8ZgxMBicQKsPfNH1PykMg16TAvrZgoxDnxr3eorcbhvZxyZzStwFkvqCJegr8Gbwj3GQum8QpXQPh7DGkoobpTB7YbcnUeUSKRDyX2cNN9h/84'/1'/0'/0/*)#ey7hlgpn");
       final receiver = BdkClient(
-          "cart super leaf clinic pistol plug replace close super tooth wealth usage");
+          "wpkh(tprv8ZgxMBicQKsPczV7D2zfMr7oUzHDhNPEuBUgrwRoWM3ijLRvhG87xYiqh9JFLPqojuhmqwMdo1oJzbe5GUpxCbDHnqyGhQa5Jg1Wt6rc9di/84'/1'/0'/0/*)#kdnuw5lq");
       await sender.restoreWallet();
       await receiver.restoreWallet();
       // Receiver creates the payjoin URI
       final pjReceiverAddress = (await receiver.getNewAddress()).address;
       final pjSenderAddress = (await sender.getNewAddress()).address;
-      await btcClient.sendToAddress(await pjSenderAddress.asString(), 10);
-      await btcClient.sendToAddress(await pjReceiverAddress.asString(), 2);
+      await btcClient.sendToAddress(await pjSenderAddress.asString(), 1);
+      await btcClient.sendToAddress(await pjReceiverAddress.asString(), 1);
       await btcClient.generate(11, await pjSenderAddress.asString());
       await receiver.syncWallet();
       await sender.syncWallet();
-      final pjUri = await payJoinLib.buildPjUri(
-          0.0083285, await pjReceiverAddress.asString());
       // Sender create a funded PSBT (not broadcast) to address with amount given in the pjUri
       debugPrint("Sender Balance: ${(await sender.getBalance()).toString()}");
-      final uri = await pay_join_uri.Uri.fromString(pjUri);
+      final uri = await pay_join_uri.Uri.fromString(
+          "${await pjReceiverAddress.toQrUri()}?amount=${0.0083285}&pj=https://example.com");
       final address = await uri.address();
       int amount = (((await uri.amount()) ?? 0) * 100000000).toInt();
+
       final senderPsbt = (await sender.createPsbt(address, amount, 2000));
       final senderPsbtBase64 = await senderPsbt.serialize();
       debugPrint(
         "\nOriginal sender psbt: $senderPsbtBase64",
       );
+
       // Receiver part
-      final (provisionalProposal, ctx) =
-          await payJoinLib.handlePjRequest(senderPsbtBase64, pjUri, (e) async {
-        final script = ScriptBuf(bytes: e);
-        return (await receiver.getAddressInfo(script));
+      final (req, ctx) = await (await (await send.RequestBuilder.fromPsbtAndUri(
+                  psbtBase64: senderPsbtBase64, uri: uri))
+              .buildWithAdditionalFee(
+                  maxFeeContribution: 10000,
+                  minFeeRate: 0,
+                  clampFeeContribution: false))
+          .extractContextV1();
+      final headers = common.Headers(map: {
+        'content-type': 'text/plain',
+        'content-length': req.body.length.toString(),
       });
-      final availableInputs = await receiver.listUnspent();
+      final uncheckedProposal = await v1.UncheckedProposal.fromRequest(
+          body: req.body.toList(),
+          query: (await req.url.query())!,
+          headers: headers);
+      // in a payment processor where the sender could go offline, this is where you schedule to broadcast the original_tx
+      var _ = await uncheckedProposal.extractTxToScheduleBroadcast();
+      final inputsOwned = await uncheckedProposal.checkBroadcastSuitability(
+          canBroadcast: (e) async {
+        return true;
+      });
+      // Receive Check 2: receiver can't sign for proposal inputs
+      final mixedInputScripts =
+          await inputsOwned.checkInputsNotOwned(isOwned: (e) async {
+        return await receiver.getAddressInfo(ScriptBuf(bytes: e));
+      });
+
+      // Receive Check 3: receiver can't sign for proposal inputs
+      final seenInputs = await mixedInputScripts.checkNoMixedInputScripts();
+      // Receive Check 4: have we seen this input before? More of a check for non-interactive i.e. payment processor receivers.
+      final provisionalProposal =
+          await (await seenInputs.checkNoInputsSeenBefore(isKnown: (e) async {
+        return false;
+      }))
+              .identifyReceiverOutputs(isReceiverOutput: (e) async {
+        return await receiver.getAddressInfo(ScriptBuf(bytes: e));
+      });
+      final unspent = await receiver.listUnspent();
       // Select receiver payjoin inputs.
       Map<int, common.OutPoint> candidateInputs = {
-        for (var input in availableInputs)
+        for (var input in unspent)
           input.txout.value: common.OutPoint(
               txid: input.outpoint.txid.toString(), vout: input.outpoint.vout)
       };
       final selectedOutpoint = await provisionalProposal.tryPreservingPrivacy(
           candidateInputs: candidateInputs);
-      var selectedUtxo = availableInputs.firstWhere(
+      var selectedUtxo = unspent.firstWhere(
           (i) =>
               i.outpoint.txid.toString() == selectedOutpoint.txid &&
               i.outpoint.vout == selectedOutpoint.vout,
@@ -81,6 +114,7 @@ void main() {
           address: await receiverAddress.asString());
       final payJoinProposal =
           await provisionalProposal.finalizeProposal(processPsbt: (e) async {
+        debugPrint("\n Original receiver unsigned psbt: $e");
         return await (await receiver
                 .signPsbt(await PartiallySignedTransaction.fromString(e)))
             .serialize();