+++ rust/src/api/receive.rs
@@ -1,4 +1,4 @@
-use crate::api::uri::Url;
+use crate::api::uri::{OhttpKeys, Url};
 use crate::frb_generated::RustOpaque;
 pub use crate::utils::error::PayjoinError;
 use crate::utils::types::{Headers, OutPoint, TxOut};
@@ -47,8 +47,20 @@ impl UncheckedProposal {
     ) -> Result<MaybeInputsOwned, PayjoinError> {
         let runtime = tokio::runtime::Runtime::new().unwrap();
         ptr.0
-            .check_broadcast_suitability_with_callback(min_fee_rate, |x| {
+            .check_broadcast_suitability(min_fee_rate, |x| {
                 Ok(runtime.block_on(can_broadcast(x.clone())))
+            })
+            .map(|e| e.into())
+            .map_err(|e| e.into())
+    }
+    pub(crate) fn _check_broadcast_suitability(
+        ptr: Self,
+        min_fee_rate: Option<u64>,
+        can_broadcast: impl Fn(Vec<u8>) -> Result<bool, PayjoinError>,
+    ) -> Result<MaybeInputsOwned, PayjoinError> {
+        ptr.0
+            .check_broadcast_suitability(min_fee_rate, |x| {
+                can_broadcast(x.clone()).map_err(|e| e.into())
             })
             .map(|e| e.into())
             .map_err(|e| e.into())
@@ -74,7 +86,16 @@ impl MaybeInputsOwned {
     ) -> Result<MaybeMixedInputScripts, PayjoinError> {
         let runtime = tokio::runtime::Runtime::new().unwrap();
         ptr.0
-            .check_inputs_not_owned_with_callback(|o| Ok(runtime.block_on(is_owned(o.clone()))))
+            .check_inputs_not_owned(|o| Ok(runtime.block_on(is_owned(o.clone()))))
+            .map(|e| e.into())
+            .map_err(|e| e.into())
+    }
+    pub(crate) fn _check_inputs_not_owned(
+        ptr: Self,
+        is_owned: impl Fn(Vec<u8>) ->Result<bool, PayjoinError>,
+    ) -> Result<MaybeMixedInputScripts, PayjoinError> {
+        ptr.0
+            .check_inputs_not_owned(|o| is_owned(o.clone()).map_err(|e| e.into()))
             .map(|e| e.into())
             .map_err(|e| e.into())
     }
@@ -110,7 +131,16 @@ impl MaybeInputsSeen {
     ) -> Result<OutputsUnknown, PayjoinError> {
         let runtime = tokio::runtime::Runtime::new().unwrap();
         ptr.0
-            .check_no_inputs_seen_before_with_callback(|o| Ok(runtime.block_on(is_known(o.into()))))
+            .check_no_inputs_seen_before(|o| Ok(runtime.block_on(is_known(o.into()))))
+            .map(|e| e.into())
+            .map_err(|e| e.into())
+    }
+   pub(crate) fn _check_no_inputs_seen_before(
+        ptr: Self,
+        is_known: impl Fn(OutPoint) -> Result<bool, PayjoinError>,
+    ) -> Result<OutputsUnknown, PayjoinError> {
+        ptr.0
+            .check_no_inputs_seen_before(|o| is_known(o.into()).map_err(|e| e.into()))
             .map(|e| e.into())
             .map_err(|e| e.into())
     }
@@ -130,8 +160,19 @@ impl OutputsUnknown {
     ) -> Result<ProvisionalProposal, PayjoinError> {
         let runtime = tokio::runtime::Runtime::new().unwrap();
         ptr.0
-            .identify_receiver_outputs_with_callback(|o| {
+            .identify_receiver_outputs(|o| {
                 Ok(runtime.block_on(is_receiver_output(o.clone())))
+            })
+            .map(|e| e.into())
+            .map_err(|e| e.into())
+    }
+    pub(crate) fn _identify_receiver_outputs(
+        ptr: Self,
+        is_receiver_output: impl Fn(Vec<u8>) -> Result<bool, PayjoinError>,
+    ) -> Result<ProvisionalProposal, PayjoinError> {
+        ptr.0
+            .identify_receiver_outputs(|o| {
+               is_receiver_output(o.clone()).map_err(|e|e.into())
             })
             .map(|e| e.into())
             .map_err(|e| e.into())
@@ -190,8 +231,21 @@ impl ProvisionalProposal {
     ) -> Result<PayjoinProposal, PayjoinError> {
         let runtime = tokio::runtime::Runtime::new().unwrap();
         ptr.0
-            .finalize_proposal_with_callback(
+            .finalize_proposal(
                 |o| Ok(runtime.block_on(process_psbt(o.clone()))),
+                min_feerate_sat_per_vb,
+            )
+            .map(|e| e.into())
+            .map_err(|e| e.into())
+    }
+    pub(crate) fn _finalize_proposal(
+        ptr: Self,
+        process_psbt: impl Fn(String) -> Result<String, PayjoinError>,
+        min_feerate_sat_per_vb: Option<u64>,
+    ) -> Result<PayjoinProposal, PayjoinError> {
+        ptr.0
+            .finalize_proposal(
+                |o| process_psbt(o.clone()).map_err(|e| e.into()),
                 min_feerate_sat_per_vb,
             )
             .map(|e| e.into())
@@ -246,28 +300,17 @@ impl From<payjoin_ffi::receive::v2::Enroller> for Enroller {
     }
 }
 impl Enroller {
-    pub fn from_relay_config(
-        relay_url: String,
-        ohttp_config_base64: String,
-        ohttp_proxy_url: String,
+    pub fn from_directory_config(
+        directory: Url, ohttp_keys: OhttpKeys, ohttp_relay:Url
     ) -> Enroller {
-        payjoin_ffi::receive::v2::Enroller::from_relay_config(
-            relay_url,
-            ohttp_config_base64,
-            ohttp_proxy_url,
-        )
-        .into()
+        payjoin_ffi::receive::v2::Enroller::from_directory_config(
+            (*directory.0).clone(), Arc::new(ohttp_keys.into()), (*ohttp_relay.0).clone()
+        ).into()
     }
 
-    pub fn subdirectory(&self) -> String {
-        self.0.subdirectory()
-    }
-    pub fn payjoin_subdir(&self) -> String {
-        self.0.payjoin_subdir()
-    }
     pub fn extract_req(ptr: Self) -> Result<((Url, Vec<u8>), ClientResponse), PayjoinError> {
         ptr.0
-            .extract_req_as_tuple()
+            .extract_req()
             .map(|e| (((*e.0.url).clone().into(), e.0.body), e.1.into()))
             .map_err(|e| e.into())
     }
@@ -277,7 +320,7 @@ impl Enroller {
         ctx: ClientResponse,
     ) -> Result<Enrolled, PayjoinError> {
         self.0
-            .process_res_with_ohttp_response(body, ctx.into())
+            .process_res(body, ctx.into())
             .map(|e| e.into())
             .map_err(|e| e.into())
     }
@@ -291,15 +334,12 @@ impl From<Arc<payjoin_ffi::receive::v2::Enrolled>> for Enrolled {
     }
 }
 impl Enrolled {
-    pub fn subdirectory(&self) -> Vec<u8> {
-        self.0.pubkey()
-    }
     pub fn fallback_target(&self) -> String {
         self.0.fallback_target()
     }
     pub fn extract_req(ptr: Self) -> Result<((Url, Vec<u8>), ClientResponse), PayjoinError> {
         ptr.0
-            .extract_req_as_tuple()
+            .extract_req()
             .map(|e| (((*e.0.url).clone().into(), e.0.body), e.1.into()))
             .map_err(|e| e.into())
     }
@@ -309,7 +349,7 @@ impl Enrolled {
         ctx: ClientResponse,
     ) -> Result<Option<V2UncheckedProposal>, PayjoinError> {
         self.0
-            .process_res_with_ohttp_response(body, ctx.into())
+            .process_res(body, ctx.into())
             .map(|e| e.map(|o| o.into()))
             .map_err(|e| e.into())
     }
@@ -341,7 +381,7 @@ impl V2UncheckedProposal {
     ) -> Result<V2MaybeInputsOwned, PayjoinError> {
         let runtime = tokio::runtime::Runtime::new().unwrap();
         self.0
-            .check_broadcast_suitability_with_callback(min_fee_rate, |x| {
+            .check_broadcast_suitability(min_fee_rate, |x| {
                 Ok(runtime.block_on(can_broadcast(x.clone())))
             })
             .map(|e| e.into())
@@ -373,7 +413,7 @@ impl V2MaybeInputsOwned {
     ) -> Result<V2MaybeMixedInputScripts, PayjoinError> {
         let runtime = tokio::runtime::Runtime::new().unwrap();
         self.0
-            .check_inputs_not_owned_with_callback(|o| Ok(runtime.block_on(is_owned(o.clone()))))
+            .check_inputs_not_owned(|o| Ok(runtime.block_on(is_owned(o.clone()))))
             .map(|e| e.into())
             .map_err(|e| e.into())
     }
@@ -419,7 +459,7 @@ impl V2MaybeInputsSeen {
     ) -> Result<V2OutputsUnknown, PayjoinError> {
         let runtime = tokio::runtime::Runtime::new().unwrap();
         self.0
-            .check_no_inputs_seen_before_with_callback(|o| Ok(runtime.block_on(is_known(o.into()))))
+            .check_no_inputs_seen_before(|o| Ok(runtime.block_on(is_known(o.into()))))
             .map(|e| e.into())
             .map_err(|e| e.into())
     }
@@ -440,7 +480,7 @@ impl V2OutputsUnknown {
     ) -> Result<V2ProvisionalProposal, PayjoinError> {
         let runtime = tokio::runtime::Runtime::new().unwrap();
         self.0
-            .identify_receiver_outputs_with_callback(|o| {
+            .identify_receiver_outputs(|o| {
                 Ok(runtime.block_on(is_receiver_output(o.clone())))
             })
             .map(|e| e.into())
@@ -505,7 +545,7 @@ impl V2ProvisionalProposal {
     ) -> Result<V2PayjoinProposal, PayjoinError> {
         let runtime = tokio::runtime::Runtime::new().unwrap();
         self.0
-            .finalize_proposal_with_callback(
+            .finalize_proposal(
                 |o| Ok(runtime.block_on(process_psbt(o.clone()))),
                 min_feerate_sat_per_vb,
             )
@@ -542,7 +582,7 @@ impl V2PayjoinProposal {
     pub fn extract_v2_req(ptr: Self) -> Result<((Url, Vec<u8>), ClientResponse), PayjoinError> {
         ptr.0
             .clone()
-            .extract_v2_req_as_tuple()
+            .extract_v2_req()
             .map(|e| (((*e.0.url).clone().into(), e.0.body), e.1.into()))
             .map_err(|e| e.into())
     }
@@ -553,7 +593,7 @@ impl V2PayjoinProposal {
         ohttp_context: ClientResponse,
     ) -> Result<Vec<u8>, PayjoinError> {
         self.0
-            .deserialize_res_with_ohttp_response(res, ohttp_context.into())
+            .deserialize_res(res, ohttp_context.into())
             .map(|e| e)
             .map_err(|e| e.into())
     }